+++ asphalt/serialization/serializers/json.py
@@ -1,11 +1,11 @@
-from typing import Dict, Any
 from json.decoder import JSONDecoder
 from json.encoder import JSONEncoder
+from typing import Dict, Any
 
-from typeguard import check_argument_types
 from asphalt.core.util import resolve_reference
+from typeguard import check_argument_types
 
-from ..api import Serializer
+from asphalt.serialization.api import Serializer
 
 
 class JSONSerializer(Serializer):
@@ -28,8 +28,8 @@ class JSONSerializer(Serializer):
 
     __slots__ = 'encoding', '_encoder', '_decoder'
 
-    def __init__(self, encoder_options: Dict[str, Any]=None, decoder_options: Dict[str, Any]=None,
-                 encoding: str='utf-8'):
+    def __init__(self, encoder_options: Dict[str, Any] = None,
+                 decoder_options: Dict[str, Any] = None, encoding: str = 'utf-8'):
         assert check_argument_types()
         self.encoding = encoding
         encoder_options = encoder_options or {}