+++ asphalt/serialization/serializers/msgpack.py
@@ -1,14 +1,9 @@
 from typing import Dict, Any
 
+from msgpack import packb, unpackb
 from typeguard import check_argument_types
 
-from ..api import Serializer
-
-try:
-    from msgpack import packb, unpackb
-except ImportError:  # pragma: no cover
-    raise ImportError('msgpack-python is missing -- install '
-                      'asphalt-serialization[msgpack] to fix this')
+from asphalt.serialization.api import Serializer
 
 
 class MsgpackSerializer(Serializer):
@@ -37,7 +32,8 @@ class MsgpackSerializer(Serializer):
 
     __slots__ = '_packer_options', '_unpacker_options'
 
-    def __init__(self, packer_options: Dict[str, Any]=None, unpacker_options: Dict[str, Any]=None):
+    def __init__(self, packer_options: Dict[str, Any] = None,
+                 unpacker_options: Dict[str, Any] = None):
         assert check_argument_types()
         packer_options = packer_options or {}
         packer_options.setdefault('use_bin_type', True)