+++ asphalt/serialization/serializers/pickle.py
@@ -2,7 +2,7 @@ import pickle
 
 from typeguard import check_argument_types
 
-from ..api import Serializer
+from asphalt.serialization.api import Serializer
 
 
 class PickleSerializer(Serializer):
@@ -17,7 +17,7 @@ class PickleSerializer(Serializer):
 
     __slots__ = 'protocol'
 
-    def __init__(self, protocol: int=pickle.HIGHEST_PROTOCOL):
+    def __init__(self, protocol: int = pickle.HIGHEST_PROTOCOL):
         assert check_argument_types()
         assert 0 <= protocol <= pickle.HIGHEST_PROTOCOL,\
             '"protocol" must be between 0 and {}'.format(pickle.HIGHEST_PROTOCOL)